+++ build.gradle
@@ -59,14 +59,8 @@ dependencies {
 	runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.11.2'
 	runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.11.2'
 
-	//MAIL
-	implementation 'org.springframework.boot:spring-boot-starter-mail'
-
-	//DEVTOOLS
+	//Devtools
 	developmentOnly 'org.springframework.boot:spring-boot-devtools'
-
-	//VALIDATOR
-	implementation 'org.springframework.boot:spring-boot-starter-validation'
 
 	//Test
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
