+++ src/main/java/com/whl/hotelService/config/auth/PrincipalDetails.java
@@ -1,6 +1,6 @@
 package com.whl.hotelService.config.auth;
 
-import com.whl.hotelService.Userdomain.dto.UserDto;
+import com.whl.hotelService.domain.dto.UserDto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -29,7 +29,6 @@ public class PrincipalDetails implements UserDetails, OAuth2User {
     public Map<String, Object> getAttributes() {
         return null;
     }
-
     @Override
     public String getName() {
         return null;
