+++ src/main/java/com/whl/hotelService/config/auth/jwt/JwtTokenProvider.java
@@ -1,7 +1,7 @@
 package com.whl.hotelService.config.auth.jwt;
 
 import com.whl.hotelService.config.auth.PrincipalDetails;
-import com.whl.hotelService.Userdomain.dto.UserDto;
+import com.whl.hotelService.domain.dto.UserDto;
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.security.Keys;
 import lombok.extern.slf4j.Slf4j;
@@ -12,10 +12,6 @@ import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Component;
 
 import java.security.Key;
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Date;
@@ -25,40 +21,13 @@ import java.util.stream.Collectors;
 @Component
 public class JwtTokenProvider {
 
-
     //Key 저장
     private final Key key;
 
-    String url  = "jdbc:mysql://localhost:3306/hotel_db";
-    String username = "root";
-    String password  = "1234";
-    Connection conn;
-    PreparedStatement pstmt;
-    ResultSet rs;
-
-    public JwtTokenProvider() throws Exception {
-
-        Class.forName("com.mysql.cj.jdbc.Driver");
-        conn = DriverManager.getConnection(url,username,password);
-        pstmt = conn.prepareStatement("select * from signature");
-        rs =pstmt.executeQuery();
-
-        if(rs.next())
-        {
-
-            byte [] keyByte =  rs.getBytes("signature");                 //DB로 서명Key꺼내옴
-            this.key = Keys.hmacShaKeyFor(keyByte);                                    //this.key에 저장
-            System.out.println("[JwtTokenProvider] Key : " + this.key );
-        }
-        else {
-            byte[] keyBytes = KeyGenerator.getKeygen();     //난수키값 가져오기
-            this.key = Keys.hmacShaKeyFor(keyBytes);        // 생성된 키를 사용하여 HMAC SHA(암호화알고리즘)알고리즘에 기반한 Key 객체 생성
-            pstmt = conn.prepareStatement("insert into signature values(?,now())");
-
-            pstmt.setBytes(1, keyBytes);
-            pstmt.executeUpdate();
-            System.out.println("[JwtTokenProvider] Constructor Key init: " + key);
-        }
+    public JwtTokenProvider() {
+        byte[] keyBytes = KeyGenerator.getKeygen();
+        this.key = Keys.hmacShaKeyFor(keyBytes);
+        System.out.println("JwtTokenProvider Constructor  Key init: " + key);
 
     }
 
@@ -104,35 +73,7 @@ public class JwtTokenProvider {
                 .build();
     }
 
-    public TokenInfo generateToken(String Claimkey,String id,boolean isAuth) {
 
-        long now = (new Date()).getTime();
-
-        // Access Token 생성
-        Date accessTokenExpiresIn = new Date(now + 60*5*1000);    // 60*5 초후 만료
-        String accessToken = Jwts.builder()
-                .setSubject(Claimkey+"JWT TOKEN")
-                .claim(Claimkey,isAuth)             //정보저장
-                .claim("id",id)
-                .setExpiration(accessTokenExpiresIn)
-                .signWith(key, SignatureAlgorithm.HS256)
-                .compact();
-
-        // Refresh Token 생성
-        String refreshToken = Jwts.builder()
-                .setExpiration(new Date(now + 86400000))    //1일: 24 * 60 * 60 * 1000 = 86400000
-                .signWith(key, SignatureAlgorithm.HS256)
-                .compact();
-
-        System.out.println("[JwtTokenProvider] generateToken() accessToken : " + accessToken);
-        System.out.println("[JwtTokenProvider] generateToken() refreshToken : " + refreshToken);
-
-        return TokenInfo.builder()
-                .grantType("Bearer")
-                .accessToken(accessToken)
-                .refreshToken(refreshToken)
-                .build();
-    }
 
 
 
@@ -179,7 +120,7 @@ public class JwtTokenProvider {
 
     }
 
-    public Claims parseClaims(String accessToken) {
+    private Claims parseClaims(String accessToken) {
         try {
             return Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(accessToken).getBody();
         } catch (ExpiredJwtException e) {
@@ -194,17 +135,14 @@ public class JwtTokenProvider {
             return true;
         } catch (io.jsonwebtoken.security.SecurityException | MalformedJwtException e) {
             log.info("Invalid JWT Token", e);
-        }
-        catch (ExpiredJwtException e) {
-            log.info("Expired JWT Token", e);
-            return false;
+//        }
+//        catch (ExpiredJwtException e) {
+//            log.info("Expired JWT Token", e);
+
         } catch (UnsupportedJwtException e) {
             log.info("Unsupported JWT Token", e);
         } catch (IllegalArgumentException e) {
             log.info("JWT claims string is empty.", e);
-        } catch(Exception etc){
-            log.info("기타예외");
-            return false;
         }
         return false;
     }
