+++ src/main/java/com/whl/hotelService/domain/dto/UserDto.java
@@ -1,0 +1,19 @@
+package com.whl.hotelService.domain.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class UserDto {
+    private String user_id;
+    private String password;
+    private String role;
+
+    private String provider;
+    private String provider_id;
+}
