+++ src/main/java/com/whl/hotelService/domain/repository/UserRepository.java
@@ -1,6 +1,7 @@
-package com.whl.hotelService.Userdomain.repository;
+package com.whl.hotelService.domain.repository;
 
-import com.whl.hotelService.Userdomain.entity.User;
+import com.whl.hotelService.domain.entity.User;
+import net.minidev.json.JSONUtil;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
