+++ src/main/resources/application.properties
@@ -42,21 +42,3 @@ spring.security.oauth2.client.provider.kakao.user-name-attribute = id
 spring.security.oauth2.client.registration.google.client-id=15984624731-jkbmoj4m3h10mkei1c0pr37q2dn97pjt.apps.googleusercontent.com
 spring.security.oauth2.client.registration.google.client-secret=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
 spring.security.oauth2.client.registration.google.scope=profile, email
-
-spring.devtools.livereload.enabled=true
-spring.devtools.restart.enabled=true
-spring.thymeleaf.cache=false
-
-#Mail Settings
-spring.mail.default-encoding=UTF-8
-
-#Google SMTP
-spring.mail.host=smtp.gmail.com
-spring.mail.port=587
-spring.mail.username=gkatjdah070@gmail.com
-spring.mail.password=awyz qmde aazq wonl
-spring.mail.properties.mail.smtp.auth=true
-spring.mail.properties.mail.debug=true
-spring.mail.properties.mail.smtp.starttls.enable=true
-spring.mail.properties.mail.mime.charset=UTF-8
-spring.mail.properties.mail.transport.protocol=smtp
