--- src/main/resources/static/user/resources/stylesheets/join.css
@@ -1,3 +1,0 @@
-form > .join_wrapper > div > .msg{
-    color : orange;
-}
